Simplify go17 and go18 differences to support http2 WrapWriter
