+++ middleware/wrap_writer_17.go
@@ -1,4 +1,4 @@
-// +build !go1.8
+// +build go1.7,!go1.8
 
 package middleware
 
@@ -9,15 +9,21 @@ import (
 
 // NewWrapResponseWriter wraps an http.ResponseWriter, returning a proxy that allows you to
 // hook into various parts of the response process.
-func NewWrapResponseWriter(w http.ResponseWriter) WrapResponseWriter {
+func NewWrapResponseWriter(w http.ResponseWriter, protoMajor int) WrapResponseWriter {
 	_, cn := w.(http.CloseNotifier)
 	_, fl := w.(http.Flusher)
-	_, hj := w.(http.Hijacker)
 	_, rf := w.(io.ReaderFrom)
 
 	bw := basicWriter{ResponseWriter: w}
-	if cn && fl && hj && rf {
-		return &httpFancyWriter{bw}
+	if protoMajor == 2 {
+		if cn && fl && rf {
+			return &http2FancyWriter{bw}
+		}
+	} else {
+		_, hj := w.(http.Hijacker)
+		if cn && fl && hj && rf {
+			return &httpFancyWriter{bw}
+		}
 	}
 	if fl {
 		return &flushWriter{bw}
